Add documentation, simplify parameters, move to separate files
